practice

+++ java-practise2.0/fundamentalProgramming/practiceQuestions/question16.java
@@ -5,18 +5,17 @@ public class question16 {
     public static void main(String[] args) {
         Scanner scn = new Scanner(System.in);
         int n = scn.nextInt();
-        int[] arr = new int[n];
+        long[] arr = new long[n];
         for(int i = 0; i < n; i++) {
-            arr[i] = scn.nextInt();
+            arr[i] = scn.nextLong();
         }
-
         productArray(arr);
     }
 
-    private static void productArray(int[] arr) {
+    private static void productArray(long[] arr) {
         //int[] arrNew = new int[arr.length];
         for (int i = 0; i < arr.length; i++) {
-            int product = 1;
+            long product = 1;
             for (int j = 0; j < arr.length; j++) {
                 if (i != j) {
                     product = product * arr[j];
